pyproject.toml
[project]
name = "personal-health-data"
version = "0.1.0"
description = "Personal health data analysis tool with interactive Dash visualization"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "dash>=3.3.0",
    "dash-bootstrap-components>=2.0.4",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scipy>=1.17.0",
]

[project.scripts]
health-data = "lib.cli:main"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
